modern architecture for android

+++ app/src/main/java/com/example/modernarchitecturesample/core/repository/MovieRepositoryImpl.kt
@@ -3,6 +3,7 @@ package com.example.modernarchitecturesample.core.repository
 import com.example.modernarchitecturesample.core.datasource.cache.CacheDataSource
 import com.example.modernarchitecturesample.core.datasource.remote.RemoteDataSource
 import com.example.modernarchitecturesample.core.repository.model.*
+import com.example.modernarchitecturesample.feature.util.NetworkUtils
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import java.util.concurrent.TimeUnit
@@ -10,7 +11,8 @@ import javax.inject.Inject
 
 class MovieRepositoryImpl @Inject constructor(
     private val remoteSource: RemoteDataSource,
-    private val cacheDataSource: CacheDataSource
+    private val cacheDataSource: CacheDataSource,
+    private val networkUtils: NetworkUtils
 ) : MovieRepository {
 
     companion object {
@@ -31,24 +33,35 @@ class MovieRepositoryImpl @Inject constructor(
     override val getCacheMovie: Flow<Results<List<Movie>>>
         get() = cacheDataSource.loadMovie().map { cacheValue ->
             if (cacheValue.isEmpty()) {
-                try {
-                    val cacheData = remoteSource.getMovie()
-                    cacheDataSource.insertMovie(*cacheData.mapListToDatabase().toTypedArray())
-                    Results.Success(cacheData.mapListMovieToRepository())
-                } catch (e: Exception) {
-                    Results.Error(e.localizedMessage ?: "Error")
-                }
-            } else {
-                if (cacheValue.first().timestamp.isExpired()) {
+                if (!networkUtils.hasNetworkConnection()) {
+                    Results.Error("Network Error")
+                } else {
                     try {
                         val cacheData = remoteSource.getMovie()
-                        cacheDataSource.deleteAllMovie()
                         cacheDataSource.insertMovie(*cacheData.mapListToDatabase().toTypedArray())
                         Results.Success(cacheData.mapListMovieToRepository())
                     } catch (e: Exception) {
                         Results.Error(e.localizedMessage ?: "Error")
                     }
-                } else Results.Success(cacheValue.mapToListMovie())
+                }
+            } else {
+                if (!networkUtils.hasNetworkConnection()) {
+                    Results.Success(cacheValue.mapToListMovie())
+                } else {
+                    if (cacheValue.first().timestamp.isExpired()) {
+                        try {
+                            val cacheData = remoteSource.getMovie()
+                            cacheDataSource.deleteAllMovie()
+                            cacheDataSource.insertMovie(
+                                *cacheData.mapListToDatabase().toTypedArray()
+                            )
+                            Results.Success(cacheData.mapListMovieToRepository())
+                        } catch (e: Exception) {
+                            Results.Error(e.localizedMessage ?: "Error")
+                        }
+                    } else Results.Success(cacheValue.mapToListMovie())
+                }
+
             }
         }
